pyproject.toml
[project]
name = "plan-migration"
version = "0.1.0"
requires-python = ">=3.13.0"
authors = [{ name = "Mark Zwart", email = "mark.zwart@pobox.com" }]
dependencies = [
    "asttokens==3.0.0",
    "blinker==1.9.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.0",
    "dash==3.0.4",
    "dash-bootstrap-components==2.0.2",
    "decorator==5.2.1",
    "et_xmlfile==2.0.0",
    "executing==2.2.0",
    "fastexcel==0.14.0",
    "Flask==3.0.3",
    "idna==3.10",
    "igraph==0.11.8",
    "importlib_metadata==8.7.0",
    "ipython==9.2.0",
    "ipython_pygments_lexers==1.1.1",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "Jinja2==3.1.6",
    "jsonpickle==4.0.5",
    "MarkupSafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "narwhals==1.39.1",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "numpy==2.2.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "plotly==6.1.0",
    "polars==1.29.0",
    "prompt_toolkit==3.0.51",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "pyarrow==20.0.0",
    "Pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyvis==0.3.2",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "retrying==1.3.4",
    "setuptools==80.7.1",
    "six==1.17.0",
    "stack-data==0.6.3",
    "texttable==1.7.0",
    "traitlets==5.14.3",
    "typing_extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "Werkzeug==3.0.6",
    "XlsxWriter==3.2.3",
    "zipp==3.21.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
